requirements.txt
backcall==0.1.0
beautifulsoup4==4.8.2
certifi==2019.11.28
chardet==3.0.4
decorator==4.4.1
idna==2.8
ipython==7.9.0
ipython-genutils==0.2.0
jedi==0.16.0
parso==0.5.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==2.0.10
ptyprocess==0.6.0
Pygments==2.5.2
PyYAML==5.3
requests==2.22.0
six==1.14.0
soupsieve==1.9.5
traitlets==4.3.3
urllib3==1.25.8
wcwidth==0.1.8
